refactor: 패키지 분리 (Price, Event)
