+++ src/main/java/kitchenpos/element/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.element;
 
 import java.math.BigDecimal;
 import java.util.Objects;
