+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -8,7 +8,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
+import kitchenpos.element.Price;
 
 @Entity
 public class Menu {
