+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -6,7 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
+import kitchenpos.element.Price;
 
 @Entity
 public class Product {
